pyproject.toml
[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
namespaces = true
where = ["src"]

[tool.setuptools.package-data]
"pyduin.data.pinfiles" = ["*.yml"]
"pyduin.data.platformio" = ["platformio.ini", "pyduin.ino"]

[project]
name = "pyduin"
version = "0.6.1"
description = "Extensive Arduino abstraction"
readme = "README.md"
authors = [
    {name = "Steffen Kockel", email = "info@steffen-kockel.de"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">3"
dependencies = [
    "platformio",
    "PyYAML",
    "pyserial",
    "requests",
    "termcolor"
]
license = {text = "GPLv3"}

[project.urls]
Homepage = "http://github.com/SteffenKockel/pyduin"
Documentation = "http://github.com/SteffenKockel/pyduin"
Source = "http://github.com/SteffenKockel/pyduin"
Tracker = "http://github.com/SteffenKockel/pyduin/issues"

[project.scripts]
pyduin = "pyduin.arduino_cli:main"

# Firmware, Pinfiles, platformio.ini